Substraction of 2 numbers using js
In this example, the parseInt() method is used to convert the user input string to a number.
The prompt() method is used to take inputs from the user. The - operator is used to subtract the second number from the first number and the result is stored in the variable difference.
Finally, the console.log() method is used to display the difference between the two numbers.