Create Multiply Two Numbers with JavaScript
In this example, the parseInt() method is used to convert the user input string to a number. The prompt() method is used to take inputs from the user. The * operator is used to multiply the two numbers and the result is stored in the variable product. Finally, the console.log() method is used to display the product of the two numbers.