Addition of 2 numbers
In this example, the parseInt() method is used to convert the user input string to a number. The prompt() method is used to take inputs from the user. The + operator is used to add the two numbers and the result is stored in the variable sum. Finally, the console.log() method is used to display the sum of the two numbers.