Finding a Remainder in JavaScript
The remainder operator in JavaScript is represented by the % symbol. It returns the remainder after the first operand is divided by the second operand. The sign of the result is the same as the sign of the first operand. Here’s an example: if we divide 10 by 3, the quotient is 3 and the remainder is 1. The modulo operation gives us only the remainder. In JavaScript, this operation is performed using the % operator.