Create Escaping Literal Quotes in Strings using js
In JavaScript, you can escape a quote from considering it as an end of string quote by placing a backslash (\) in front of the quote. For example, if you want to use a double quote inside a string that is wrapped in double quotes, you would terminate the string prematurely. To avoid this, you can use the backslash character to escape the double quote and treat it as a literal double quote character instead of an end-of-string character. Here is an example: