Escape Sequences in Strings in js
In JavaScript, escape sequences are used to represent certain characters that cannot be typed directly in a string. The backslash character (\) is used to indicate the beginning of an escape sequence. Here are some commonly used escape sequences in JavaScript:

\': Single quote
\": Double quote
\\: Backslash
\n: Newline
\r: Carriage return
\t: Tab
\b: Backspace
\f: Form feed
For example, if you want to include a double quote inside a string that is wrapped in double quotes, you can use the escape sequence \" to represent the double quote character. Here is an example: