Create Use Bracket Notation to Find the Last Character in a String using js
Here, str.length returns the length of the string, and str.length - 1 gives the index of the last character in the string. We then use this index to access the last character of the string using bracket notation.

Alternatively, we can also use the charAt() method to get the last character of a string. Here’s how: