Create Use Bracket Notation to Find the Nth-to-Last Character in a String using js
To find the Nth-to-Last character in a string using JavaScript, you can use the bracket notation. The bracket notation is used to access a specific character in a string by its index. The index of the first character in a string is 0, and the index of the last character is the length of the string minus 1. To find the Nth-to-Last character in a string, you can subtract N from the length of the string and then use the bracket notation to access the character at that index. Here is an example: