Create Store Multiple Values in one Variable using JavaScript Arrays
In JavaScript, we can store multiple values in one variable using arrays. An array is a collection of elements that can be of different data types such as numbers, strings, objects, and even other arrays. We can declare an array using square brackets [] and separate each element with a comma. Here’s an example: